requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Flask == 2.1.2
Pillow == 8.4.0
PyYAML == 6.0
Werkzeug == 2.1.2
abl-quantconnect-stubs == 13987.1
alauda-pytest == 5.2.3
asynctest == 0.13.0
atheris == 2.0.11
base-atlas == 1.2.1
basemap == 1.3.3
basemap-data == 1.3.2
basemap-data-hires == 1.3.2
distributed == 2022.6.1
dummy-wx == 0.3.0
easydialogs-gtk == 0.1
fairmongo == 1.2.0
fqdn == 1.5.1
freetype == Milestone1
fs == 2.4.16
genapi == 0.0.8
genshi == 0.7.7
gunicorn == 20.1.0
hid == 1.0.5
httpsproxy-urllib2 == 1.0
hypothesis == 6.47.5
idna-ssl == 1.1.0
ignition-api == 8.1.18
import-java == 0.6
importlib-resources == 5.8.0
isoduration == 20.11.0
jpype1 == 1.4.0
jsonpointer == 2.3
magicofinputandoutputbykeshari == 0.0.1
magicprintbykeshari == 0.0.1
manifest-bouncer == 0.1.1
matplotlib == 3.5.2
mpl-toolkits-clifford == 0.0.3
mrlpy == 1.0.0
numba == 0.55.2
numpy == 1.23.0
olefile == 0.46
openbaton-cli == 6.0.0
opencv_python == 4.6.0.66
org-geppetto-recording == 0.1.0
os-sys == 2.1.4
os-sys-linux == 0.0.0.dev0
printshortcut == 0.0.2
py4grid == 1.0.5
pybind11 == 2.9.2
pyinstaller == 5.1
pyinstallercustom == 3.5
pyobjc-framework-cocoa == 8.5
pyobjc-framework-quartz == 8.5
pyopenssl == 22.0.0
pyperf == 2.3.1
pyqt5-sip == 12.11.0
pyqt6-sip == 13.4.0
pyrsistent == 0.18.1
pyside2 == 5.15.2.1
pyside6 == 6.3.1
pyside6-addons == 6.3.1
pyside6-essentials == 6.3.1
pytest == 7.1.2
python-ntlm == 1.1.0
python-ntlm-fixed == 1.0
python-plugin-sdk == 6.0.1
pythran == 0.11.0
pywin32 == 304
quantconnect-stubs == 14150
railroad == 0.5.0
reportlab == 3.6.10
rfc3339-validator == 0.1.4
rfc3986-validator == 0.1.1
selector-standardization-beam == 0.4.15
selector-standardizers == 0.8.3
setuptools == 62.3.4
shortcutio == 0.0.1
shortcutprint == 0.0.1
six == 1.16.0
skia-pathops == 0.7.2
skimage == 0.0
soloman == 3.1.1
soloman-controls == 3.0
sphinx == 5.0.2
toml == 0.10.2
tornado == 6.1
uri-template == 1.2.0
vampireapi == 0.3.1
watchdog == 2.1.9
webcolors == 1.12
win32compat == 221.26
wxpython == 4.1.1
wxpython-zombie == 3.1.5.5
wxpython310 == 4.1.2a2
zopfli == 0.2.1
